feat(web): show what a person's voiceprint was trained on
Splits PersonEditor into a tab shell plus PersonProfileTab (moved unchanged),
and adds a Voiceprint tab listing the contributing recordings - data that has
been on the API since voiceprints shipped and was rendered nowhere.

Panels are hidden rather than unmounted so switching tabs cannot discard a
half-typed edit; the voiceprint tab mounts on first open so the directory does
not fetch training data for every row you click through.
