+++ apps/web/src/locales/de/people.json
@@ -30,6 +30,15 @@
   "hasVoiceprint": "Hat einen Stimmabdruck",
   "noVoiceprint": "Kein Stimmabdruck",
   "voiceprintShort": "Stimme",
+  "tabProfile": "Profil",
+  "tabVoiceprint": "Stimmabdruck",
+  "voiceprintTrainedFrom": "Aus {{count}} Aufnahme trainiert",
+  "voiceprintTrainedFrom_other": "Aus {{count}} Aufnahmen trainiert",
+  "voiceprintNoSamples": "Noch keine Aufnahmen trainieren diesen Stimmabdruck.",
+  "voiceprintUsingOf": "Verwendet {{used}} von {{selected}} ausgewaehlt",
+  "voiceprintRecomputing": "Wird neu berechnet...",
+  "voiceprintNeedsRecompute": "Neuberechnung noetig",
+  "voiceprintOptedOut": "Diese Person hat den Stimmabdruck abgelehnt, es gibt nichts auszuwaehlen.",
   "identitySelf": "{{email}} - Ihr Konto",
   "identitySelfNoEmail": "Ihr Konto",
   "identityLinked": "{{email}} - Diariz-Konto",
