+++ apps/web/src/locales/en/people.json
@@ -30,6 +30,15 @@
   "hasVoiceprint": "Has a voiceprint",
   "noVoiceprint": "No voiceprint",
   "voiceprintShort": "voice",
+  "tabProfile": "Profile",
+  "tabVoiceprint": "Voiceprint",
+  "voiceprintTrainedFrom": "Trained from {{count}} recording",
+  "voiceprintTrainedFrom_other": "Trained from {{count}} recordings",
+  "voiceprintNoSamples": "No recordings train this voiceprint yet.",
+  "voiceprintUsingOf": "Using {{used}} of the {{selected}} selected",
+  "voiceprintRecomputing": "Recomputing...",
+  "voiceprintNeedsRecompute": "Needs recomputing",
+  "voiceprintOptedOut": "This person has opted out of voice-printing, so there is nothing to select.",
   "identitySelf": "{{email}} - your account",
   "identitySelfNoEmail": "your account",
   "identityLinked": "{{email}} - Diariz account",
