+++ apps/web/src/locales/es/people.json
@@ -30,6 +30,15 @@
   "hasVoiceprint": "Tiene huella de voz",
   "noVoiceprint": "Sin huella de voz",
   "voiceprintShort": "voz",
+  "tabProfile": "Perfil",
+  "tabVoiceprint": "Huella de voz",
+  "voiceprintTrainedFrom": "Entrenada con {{count}} grabacion",
+  "voiceprintTrainedFrom_other": "Entrenada con {{count}} grabaciones",
+  "voiceprintNoSamples": "Todavia no hay grabaciones que entrenen esta huella.",
+  "voiceprintUsingOf": "Usando {{used}} de los {{selected}} seleccionados",
+  "voiceprintRecomputing": "Recalculando...",
+  "voiceprintNeedsRecompute": "Hay que recalcular",
+  "voiceprintOptedOut": "Esta persona ha rechazado la huella de voz, no hay nada que seleccionar.",
   "identitySelf": "{{email}} - tu cuenta",
   "identitySelfNoEmail": "tu cuenta",
   "identityLinked": "{{email}} - cuenta de Diariz",
