+++ apps/web/src/locales/fr/people.json
@@ -30,6 +30,15 @@
   "hasVoiceprint": "Possede une empreinte vocale",
   "noVoiceprint": "Aucune empreinte vocale",
   "voiceprintShort": "voix",
+  "tabProfile": "Profil",
+  "tabVoiceprint": "Empreinte vocale",
+  "voiceprintTrainedFrom": "Entrainee a partir de {{count}} enregistrement",
+  "voiceprintTrainedFrom_other": "Entrainee a partir de {{count}} enregistrements",
+  "voiceprintNoSamples": "Aucun enregistrement n'entraine encore cette empreinte.",
+  "voiceprintUsingOf": "Utilise {{used}} sur les {{selected}} selectionnes",
+  "voiceprintRecomputing": "Recalcul en cours...",
+  "voiceprintNeedsRecompute": "A recalculer",
+  "voiceprintOptedOut": "Cette personne a refuse l'empreinte vocale, il n'y a rien a selectionner.",
   "identitySelf": "{{email}} - votre compte",
   "identitySelfNoEmail": "votre compte",
   "identityLinked": "{{email}} - compte Diariz",
